requirements.txt
streamlit
psycopg2-binary
python-dotenv
web3
eth-utils
pandas
sqlalchemy
plotly
altair
